requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==3.2.25
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
idna==3.10
pycparser==2.21
PyJWT==1.7.1
pytz==2025.2
requests==2.31.0
split-settings==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7